requirements.txt
python-dotenv
langchain
torch
langchain-huggingface
langchain_mlx
mlx-lm